Extract nav test, rename nav partial
I moved the test for the navigation partial to its own test within a new
directory called partials. I like this better for keeping the tests of
the different partials I will be using separate. Otherwise, all my tests
for the different partials that show on the homepage would be stuffed
into the page_controller_test.

I also renamed the navigation partial from app_nav to _navigation. I
prefer this as it makes it a bit more obvious that it's a partial and
not its own page.