Add blog tests and partial, move partials
I created the blog sections tests and partial. In doing so, I decided
that I did not like rendering the partials inside app.html.eex. They
should instead be rendered inside index.html.eex. I moved the rendering
of the partials to index.html and put that back to being rendered in
app.html as it was originally.