Fix test by passing full URL (not a good solution)
After adding the subdomain router yesterday, realized today that all of
my tests are broken. After looking into it a bit more, the tests are
being sent to www.example.com. Since www is a subdomain, this is sending
all of my test requests to the wrong router causing them to fail. I can
get the test to pass if I manually pass in the full address of
http://localhost:4000 and that is what I did for this test. However,
this is not a good solution as it does not fix the issue. Going to have
to look into this more, namely, catching the subdomains of www and
sending them through the regular router.