Turn into umbrella app and add Tilex (#3)
* Move main app in prep to create umbrella app

* Exclude apps/tilex

I turned my personal site into an umbrella to include Hashrocket's Tilex
open source project. It has its own git repo that it is connected to so
I am keeping it out of the repo for my personal site.

* Add mix and config files for the umbrella

* Remove path lines

* Add tilex to dependencies

* Renamed child app to tbe_elixir

This was necessary as otherwise the app was not found. When starting the
server, it would say the app tbe-elixir could not be found. After
renaming the folder from tbe-elixir to tbe_elixir, I was able to start
up the server.

* Remove unneeded paths

* Update plug so til sub goes to Tilex router

* Change plug.static to pull from Tilex

Tilex has more complicated css and js. Changed plug.static to use
Tilex's css and js as it will be easier to add in the simpler css and
js from my personal site.

* Add missing halt to avoid error on page load

* Change unused variable name to remove warning

* Rename main to avoid css conflicts