Update plug and add tilex as dependency
Heroku logs are stating:
(ArgumentError) unknown application: :tilex

I am adjusting the plug as per master_proxy to send calls to sub til to
the TilexWeb.Endpoint instead of the router. I am also adding Tilex as a
dependency to tbe_elixir. Not sure if this will get it workin on Heroku
but giving it a shot. One of many attempts I am sure.