Add tilex to deps
I've been keeping tilex as a separate submodule in git. It seemed like a
good idea at the time to keep changes separate and making it simple to
update with any changes Hashrocket makes to it in the future. However,
this *may* have been my downfall. Everything in the tilex folder is not
being uploaded to github, which means it is not being uploaded to
Heroku. I think I've been running my self in circles because of it.