Remove unneeded code
Removed some of the unneeded code in the case statement for my plug. I
realized that having anyting other than the til and catchall was
unnecessary.

I moved the tilrouter from the plug call in my endpoint. I also need to
go through and remove the router and its controller and pages as they
are not needed any longer.