Fix how compiled assets are loaded
Why this change is necessary:

* Both apps in the umbrella were using app.css and app.js for the names
for their compiled files. This caused an issue with only loading the
assets for the main app and sharing it across the entire umbrella. This
also caused me to use a janky solution to get things working when I did
not properly understand the problem.

How it addresses the issue:

* By renaming the compiled asset files for Tile to tilex.css and
tilex.js, I was able to remove the janky solution I had previously
implemented. This allowed both apps to have access to their correct css
and js files.